requirements.txt
numpy
azure-functions
flask
flask_cors
python-dotenv==1.0.0
requests==2.32.3
requests-oauthlib==2.0.0
pyodbc
bcrypt 
Flask-Mail
langchain
openai
azure-ai-documentintelligence
azure-ai-formrecognizer
Pillow
langchain-community